feat(visitor-link-usage-middleware): add a middleware to handle visitor link usage and limit
Previously, each visitor link had an unlimited number of usages. Now we introduce the
VisitorCountMiddleware middleware which is responsible for counting and limiting visitor link usage.
In details, it firstly checks whether the request comes from a Visitor (either a new one or a
returning one). If not, it passes the request as it is to the next middleware. Then, it increments a
counter (request.visitor.usage_count += 1) that measures the number of visits for the link. Finally,
it marks the link (visitor) as inactive (request.user.is_visitor = False and
request.visitor.is_active = False), if the number of visits has exceeded the maximum number allowed
originally and persists any changes to the request.visitor (field 'usage_count' and potentially
'is_active') in the database.